+++ experimental/tools/unit-generators/vhdl/generators/handshake/fork.py
@@ -1,5 +1,4 @@
-from generators.support.signal_manager import generate_signal_manager
-from generators.support.utils import get_concat_extra_signals_bitwidth
+from generators.support.signal_manager import generate_signal_manager, get_concat_extra_signals_bitwidth
 from generators.support.logic import generate_or_n
 from generators.support.eager_fork_register_block import generate_eager_fork_register_block
