+++ experimental/tools/unit-generators/vhdl/generators/handshake/ofifo.py
@@ -1,5 +1,4 @@
-from generators.support.signal_manager import generate_signal_manager
-from generators.support.utils import get_concat_extra_signals_bitwidth
+from generators.support.signal_manager import generate_signal_manager, get_concat_extra_signals_bitwidth
 from generators.handshake.tehb import generate_tehb
 from generators.support.elastic_fifo_inner import generate_elastic_fifo_inner
