+++ experimental/tools/unit-generators/vhdl/generators/handshake/tehb.py
@@ -1,5 +1,4 @@
-from generators.support.signal_manager import generate_signal_manager
-from generators.support.utils import get_concat_extra_signals_bitwidth
+from generators.support.signal_manager import generate_signal_manager, get_concat_extra_signals_bitwidth
 
 
 def generate_tehb(name, params):
